requirements.txt
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
